requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
tavily-python>=0.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
httpx>=0.24.0
